+++ NEWS.md
@@ -1,6 +1,7 @@
 # prettifyAddins 1.0.0
 
-Added non-interactive versions of the addins, thanks to the V8 package.
+* Added non-interactive versions of the addins, thanks to the V8 package.
+* The number of spaces for the indentation is now based on the RStudio preference.
 
 
 # prettifyAddins 0.1.0